Move ExecutionGraph encoding and decoding logic into execution_graph for better encapsulation
