+++ ballista/rust/scheduler/src/state/task_manager.rs
@@ -18,30 +18,29 @@
 use crate::scheduler_server::event::QueryStageSchedulerEvent;
 use crate::scheduler_server::SessionBuilder;
 use crate::state::backend::{Keyspace, StateBackendClient};
-use crate::state::execution_graph::{ExecutionGraph, ExecutionStage, StageOutput, Task};
+use crate::state::execution_graph::{ExecutionGraph, Task};
 use crate::state::executor_manager::ExecutorReservation;
 use crate::state::{decode_protobuf, encode_protobuf, with_lock};
 use ballista_core::config::BallistaConfig;
-use ballista_core::error::{BallistaError, Result};
-use ballista_core::serde::physical_plan::from_proto::parse_protobuf_hash_partitioning;
+#[cfg(not(test))]
+use ballista_core::error::BallistaError;
+use ballista_core::error::Result;
 use ballista_core::serde::protobuf::executor_grpc_client::ExecutorGrpcClient;
 
 use crate::state::session_manager::create_datafusion_context;
 use ballista_core::serde::protobuf::{
-    self, job_status, task_status, FailedJob, JobStatus, PartitionId, TaskDefinition,
-    TaskStatus,
+    self, job_status, FailedJob, JobStatus, PartitionId, TaskDefinition, TaskStatus,
 };
 use ballista_core::serde::scheduler::to_proto::hash_partitioning_to_proto;
-use ballista_core::serde::scheduler::{ExecutorMetadata, PartitionLocation};
+use ballista_core::serde::scheduler::ExecutorMetadata;
 use ballista_core::serde::{AsExecutionPlan, BallistaCodec};
-use datafusion::physical_plan::{ExecutionPlan, Partitioning};
+use datafusion::physical_plan::ExecutionPlan;
 use datafusion::prelude::SessionContext;
 use datafusion_proto::logical_plan::AsLogicalPlan;
 use log::{debug, info, warn};
 use rand::distributions::Alphanumeric;
 use rand::{thread_rng, Rng};
 use std::collections::{HashMap, HashSet};
-use std::convert::TryInto;
 use std::default::Default;
 use std::sync::Arc;
 use tokio::sync::RwLock;
@@ -484,206 +483,14 @@ impl<T: 'static + AsLogicalPlan, U: 'static + AsExecutionPlan> TaskManager<T, U>
         let session_id = &proto.session_id;
 
         let session_ctx = self.get_session(session_id).await?;
-        let mut stages: HashMap<usize, ExecutionStage> = HashMap::new();
-        for stage in proto.stages {
-            let plan_proto = U::try_decode(stage.plan.as_slice())?;
-            let plan = plan_proto.try_into_physical_plan(
-                session_ctx.as_ref(),
-                session_ctx.runtime_env().as_ref(),
-                self.codec.physical_extension_codec(),
-            )?;
 
-            let stage_id = stage.stage_id as usize;
-            let partitions: usize = stage.partitions as usize;
-
-            let mut task_statuses: Vec<Option<task_status::Status>> =
-                vec![None; partitions];
-
-            for status in stage.task_statuses {
-                if let Some(task_id) = status.task_id.as_ref() {
-                    task_statuses[task_id.partition_id as usize] = status.status
-                }
-            }
-
-            let output_partitioning: Option<Partitioning> =
-                parse_protobuf_hash_partitioning(
-                    stage.output_partitioning.as_ref(),
-                    session_ctx.as_ref(),
-                    plan.schema().as_ref(),
-                )?;
-
-            let mut inputs: HashMap<usize, StageOutput> = HashMap::new();
-
-            for input in stage.inputs {
-                let stage_id = input.stage_id as usize;
-
-                let outputs = input
-                    .partition_locations
-                    .into_iter()
-                    .map(|loc| {
-                        let partition = loc.partition as usize;
-                        let locations = loc
-                            .partition_location
-                            .into_iter()
-                            .map(|l| l.try_into())
-                            .collect::<Result<Vec<_>>>()?;
-                        Ok((partition, locations))
-                    })
-                    .collect::<Result<HashMap<usize, Vec<PartitionLocation>>>>()?;
-
-                inputs.insert(
-                    stage_id,
-                    StageOutput {
-                        partition_locations: outputs,
-                        complete: input.complete,
-                    },
-                );
-            }
-            let stage_metrics = if stage.stage_metrics.is_empty() {
-                None
-            } else {
-                let ms = stage
-                    .stage_metrics
-                    .into_iter()
-                    .map(|m| m.try_into())
-                    .collect::<Result<Vec<_>>>()?;
-                Some(ms)
-            };
-
-            let execution_stage = ExecutionStage {
-                stage_id: stage.stage_id as usize,
-                partitions,
-                output_partitioning,
-                inputs,
-                plan,
-                task_statuses,
-                output_links: stage
-                    .output_links
-                    .into_iter()
-                    .map(|l| l as usize)
-                    .collect(),
-                resolved: stage.resolved,
-                stage_metrics,
-            };
-            stages.insert(stage_id, execution_stage);
-        }
-
-        let output_locations: Vec<PartitionLocation> = proto
-            .output_locations
-            .into_iter()
-            .map(|loc| loc.try_into())
-            .collect::<Result<Vec<_>>>()?;
-
-        Ok(ExecutionGraph {
-            job_id: proto.job_id,
-            session_id: proto.session_id,
-            status: proto.status.ok_or_else(|| {
-                BallistaError::Internal(
-                    "Invalid Execution Graph: missing job status".to_owned(),
-                )
-            })?,
-            stages,
-            output_partitions: proto.output_partitions as usize,
-            output_locations,
-        })
+        ExecutionGraph::decode_execution_graph(proto, &self.codec, &session_ctx).await
     }
 
     fn encode_execution_graph(&self, graph: ExecutionGraph) -> Result<Vec<u8>> {
-        let job_id = graph.job_id().to_owned();
+        let proto = ExecutionGraph::encode_execution_graph(graph, &self.codec)?;
 
-        let stages = graph
-            .stages
-            .into_iter()
-            .map(|(stage_id, stage)| {
-                let mut plan: Vec<u8> = vec![];
-
-                U::try_from_physical_plan(
-                    stage.plan,
-                    self.codec.physical_extension_codec(),
-                )
-                .and_then(|proto| proto.try_encode(&mut plan))?;
-
-                let mut inputs: Vec<protobuf::GraphStageInput> = vec![];
-
-                for (stage, output) in stage.inputs.into_iter() {
-                    inputs.push(protobuf::GraphStageInput {
-                        stage_id: stage as u32,
-                        partition_locations: output
-                            .partition_locations
-                            .into_iter()
-                            .map(|(partition, locations)| {
-                                Ok(protobuf::TaskInputPartitions {
-                                    partition: partition as u32,
-                                    partition_location: locations
-                                        .into_iter()
-                                        .map(|l| l.try_into())
-                                        .collect::<Result<Vec<_>>>()?,
-                                })
-                            })
-                            .collect::<Result<Vec<_>>>()?,
-                        complete: output.complete,
-                    });
-                }
-
-                let task_statuses: Vec<protobuf::TaskStatus> = stage
-                    .task_statuses
-                    .into_iter()
-                    .enumerate()
-                    .filter_map(|(partition, status)| {
-                        status.map(|status| protobuf::TaskStatus {
-                            task_id: Some(protobuf::PartitionId {
-                                job_id: job_id.clone(),
-                                stage_id: stage_id as u32,
-                                partition_id: partition as u32,
-                            }),
-                            // task metrics should not persist.
-                            metrics: vec![],
-                            status: Some(status),
-                        })
-                    })
-                    .collect();
-
-                let output_partitioning =
-                    hash_partitioning_to_proto(stage.output_partitioning.as_ref())?;
-
-                let stage_metrics = stage
-                    .stage_metrics
-                    .unwrap_or_default()
-                    .into_iter()
-                    .map(|m| m.try_into())
-                    .collect::<Result<Vec<_>>>()?;
-                Ok(protobuf::ExecutionGraphStage {
-                    stage_id: stage_id as u64,
-                    partitions: stage.partitions as u32,
-                    output_partitioning,
-                    inputs,
-                    plan,
-                    task_statuses,
-                    output_links: stage
-                        .output_links
-                        .into_iter()
-                        .map(|l| l as u32)
-                        .collect(),
-                    resolved: stage.resolved,
-                    stage_metrics,
-                })
-            })
-            .collect::<Result<Vec<_>>>()?;
-
-        let output_locations: Vec<protobuf::PartitionLocation> = graph
-            .output_locations
-            .into_iter()
-            .map(|loc| loc.try_into())
-            .collect::<Result<Vec<_>>>()?;
-
-        encode_protobuf(&protobuf::ExecutionGraph {
-            job_id: graph.job_id,
-            session_id: graph.session_id,
-            status: Some(graph.status),
-            stages,
-            output_partitions: graph.output_partitions as u64,
-            output_locations,
-        })
+        encode_protobuf(&proto)
     }
 }
